+++ Blocks5/src/toxicgas.cpp
@@ -71,7 +71,7 @@ void ToxicGas::onUpdate()
 		p.sizeOnTexture = Vec2b(16, 16);
 		p.position = position * 16 + Vec2i(random(2, 14), random(2, 14));
 		const float r = random(0.0f, 6.283f);
-		p.velocity = random(0.0f, 1.0f) * Vec2f(sin(r), cos(r));
+		p.velocity = random(0.0f, 1.0f) * Vec2f(sinf(r), cosf(r));
 		p.color = Vec4f(random(0.4f, 1.0f), random(0.75f, 1.0f), random(0.0f, 0.5f), random(0.5f, 1.5f));
 		p.deltaColor = Vec4f(0.0f, 0.0f, 0.0f, -p.color.a / p.lifetime);
 		p.rotation = random(0.0f, 10.0f);
